% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozcran.R
\name{oz_db}
\alias{oz_db}
\title{Australian CRAN mirrors}
\usage{
oz_db(...)
}
\arguments{
\item{...}{dots, ignored}
}
\value{
data frame of packages from each repository
}
\description{
Database of packages from Australian CRAN mirrors
}
\details{
The data returned is from [tools::CRAN_package_db()] and not [utils::available.packages()], but with the
duplicated "MD5Sums" column removed (which breaks tibble/dplyr.) There is an added column `repos` which
is the name for the address provided above.
}
\examples{
library(dplyr)
oz_db() \%>\% group_by(repos) \%>\%
summarize(n = n(), date = max(as.Date(Published), na.rm = TRUE)) \%>\%
arrange(desc(date))
}
\seealso{
[oz_repos()] for the list of mirrors used
}
