% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozcran.R
\name{oz_repos}
\alias{oz_repos}
\title{Australian CRAN mirrors}
\usage{
oz_repos()
}
\value{
named vector of mirror addresses
}
\description{
Hardcoded list of Australian mirrors
}
\details{
These are by name \href{https://cran.csiro.au}{csiro}, \href{https://cran.ms.unimelb.edu.au}{unimelb},
\href{https://mirror.aarnet.edu.au/pub/CRAN}{aarnet}, \href{https://cran.curtin.edu.au}{curtin},
\href{https://cloud.r-project.org}{aws}.

\code{aws} is a special case, the "cloud" redirect sponsored by RStudio, it redirects to a nearby Amazon
AWS server. I asumme this is usually Sydney, right on the big fat fibre pipes in Haymarket that many
companies use.

A 'private' repository can be included by setting the environment variable 'OZ_PRIVATE'. Do not include
the trailing slash, as in the list above. If set this will be included like
\code{c(private = "https://mycran/")}.
}
\examples{
oz_repos()
}
\seealso{
\link{oz_db} that returns the package list from these repositories
}
